+++ src/Dax.Vpax.CLI/Commands/Export/ExportCommandHandler.cs
@@ -1,6 +1,6 @@
-﻿using static Dax.Vpax.CLI.Commands.ExportCommandOptions;
+﻿using static Dax.Vpax.CLI.Commands.Export.ExportCommandOptions;
 
-namespace Dax.Vpax.CLI.Commands;
+namespace Dax.Vpax.CLI.Commands.Export;
 
 internal sealed class ExportCommandHandler : ICommandHandler
 {
